refactor: Implement for interface for scancode file
Note: Added a converter class for scancode to
Opossum to allow for individual testing
