+++ src/opossum_lib/core/input_format_reader.py
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: TNG Technology Consulting GmbH <https://www.tngtech.com>
+#
+# SPDX-License-Identifier: Apache-2.0
+
 from abc import abstractmethod
 from typing import Protocol
