+++ src/opossum_lib/core/opossum_generator.py
@@ -10,14 +10,13 @@ from opossum_lib.core.opossum_model import Opossum
 from opossum_lib.opossum.file_generation import OpossumFileWriter
 from opossum_lib.opossum.opossum_file_content import OpossumFileContent
 from opossum_lib.opossum.opossum_format_reader import OpossumFormatReader
-from opossum_lib.scancode.convert_scancode_to_opossum import (
-    convert_scancode_file_to_opossum,
-)
+from opossum_lib.scancode.scancode_format_reader import ScancodeFormatReader
 from opossum_lib.spdx.convert_to_opossum import convert_spdx_to_opossum_information
 
 
 class OpossumGenerator:
     opossum_format_reader: OpossumFormatReader = OpossumFormatReader()
+    scancode_format_reader: ScancodeFormatReader = ScancodeFormatReader()
 
     def generate(
         self, opossum_generation_arguments: OpossumGenerationArguments
@@ -43,8 +42,8 @@ class OpossumGenerator:
             return self._read_to_internal_format(input_file).to_opossum_file_format()
 
     def _read_to_internal_format(self, input_file: InputFile) -> Opossum:
-        if input_file.type == FileType.SCAN_CODE:
-            return convert_scancode_file_to_opossum(input_file.path)
+        if self.scancode_format_reader.can_handle(input_file.type):
+            return self.scancode_format_reader.read(input_file.path)
         elif self.opossum_format_reader.can_handle(input_file.type):
             return self.opossum_format_reader.read(input_file.path)
         else:
