+++ src/opossum_lib/scancode/scancode_data_to_opossum_converter.py
@@ -1,0 +1,96 @@
+# SPDX-FileCopyrightText: TNG Technology Consulting GmbH <https://www.tngtech.com>
+#
+# SPDX-License-Identifier: Apache-2.0
+
+import logging
+import sys
+import uuid
+from pathlib import PurePath
+
+from opossum_lib.core.opossum_model import (
+    Metadata,
+    Opossum,
+    OpossumPackage,
+    Resource,
+    ResourceType,
+    ScanResults,
+    SourceInfo,
+)
+from opossum_lib.scancode.constants import SCANCODE_SOURCE_NAME
+from opossum_lib.scancode.model import File, FileType, Header, ScanCodeData
+
+
+class ScancodeDataToOpossumConverter:
+    @staticmethod
+    def convert_scancode_to_opossum(scancode_data: ScanCodeData) -> Opossum:
+        resources = ScancodeDataToOpossumConverter.extract_opossum_resources(
+            scancode_data
+        )
+
+        scancode_header = ScancodeDataToOpossumConverter.extract_scancode_header(
+            scancode_data
+        )
+        metadata = Metadata(
+            project_id=str(uuid.uuid4()),
+            file_creation_date=scancode_header.end_timestamp,
+            project_title="ScanCode file",
+        )
+
+        return Opossum(
+            scan_results=ScanResults(
+                metadata=metadata,
+                resources=resources,
+            )
+        )
+
+    @staticmethod
+    def extract_scancode_header(scancode_data: ScanCodeData) -> Header:
+        if len(scancode_data.headers) != 1:
+            logging.error("Headers of ScanCode file are invalid.")
+            sys.exit(1)
+        return scancode_data.headers[0]
+
+    @staticmethod
+    def extract_opossum_resources(
+        scancode_data: ScanCodeData,
+    ) -> list[Resource]:
+        temp_root = Resource(path=PurePath(""))
+        for file in scancode_data.files:
+            resource = Resource(
+                path=PurePath(file.path),
+                attributions=ScancodeDataToOpossumConverter.get_attribution_info(file),
+                type=ScancodeDataToOpossumConverter.convert_resource_type(file.type),
+            )
+            temp_root.add_resource(resource)
+
+        return list(temp_root.children.values())
+
+    @staticmethod
+    def convert_resource_type(file_type: FileType) -> ResourceType:
+        if file_type == FileType.FILE:
+            return ResourceType.FILE
+        else:
+            return ResourceType.FOLDER
+
+    @staticmethod
+    def get_attribution_info(file: File) -> list[OpossumPackage]:
+        if file.type == FileType.DIRECTORY:
+            return []
+        copyright = "\n".join(c.copyright for c in file.copyrights)
+        source_info = SourceInfo(name=SCANCODE_SOURCE_NAME)
+
+        attribution_infos = []
+        for license_detection in file.license_detections:
+            license_name = license_detection.license_expression_spdx
+            max_score = max(m.score for m in license_detection.matches)
+            attribution_confidence = int(max_score)
+
+            package = OpossumPackage(
+                source=source_info,
+                license_name=license_name,
+                attribution_confidence=attribution_confidence,
+                copyright=copyright,
+            )
+            attribution_infos.append(package)
+
+        return attribution_infos
